blog_api/requirements.txt
boto3
python-slugify
streaming-form-data
pillow~=9.5.0
markdown~=3.4.3
bleach~=6.0.0